pyproject.toml
[project]
name = "sova-parser"
version = "0.1.0"
description = "Service for parse and processing popular types documents"
authors = [
    {name = "v666k0"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "uvicorn[standart] (>=0.34.2,<0.35.0)",
    "fastapi[standart] (>=0.115.12,<0.116.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "docling (>=2.36.1,<3.0.0)",
    "ultralytics (>=8.3.152,<9.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "docling-core (>=2.44.2,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
